refactor: breakdown register form and add general improvements
