+++ apps/mobile/src/components/register-adoption-form/adoption-form.hooks.ts
@@ -1,16 +1,16 @@
 import { useState } from "react";
 import { stepsLibrary } from "./adoption-form.constants";
-import { Step, Steps } from "./adoption-form.types";
+import { Step, AdoptionSteps } from "./adoption-form.types";
 
-export const getStepsByOrder = (stepsLibrary: { [key in Steps]: Step }): {
-  [key: number]: Steps;
+export const getStepsByOrder = (stepsLibrary: { [key in AdoptionSteps]: Step }): {
+  [key: number]: AdoptionSteps;
 } =>
-  Object.keys(Steps).reduce((prev, stepId) => {
-    const step = stepId as Steps;
+  Object.keys(AdoptionSteps).reduce((prev, stepId) => {
+    const step = stepId as AdoptionSteps;
     return { ...prev, [stepsLibrary[step].order]: step };
   }, {});
 
-export function useMultiStepNavigation(initialStep = Steps.PetName) {
+export function useMultiStepNavigation(initialStep = AdoptionSteps.PetName) {
   const [activeStep, setActiveStep] = useState(initialStep);
 
   const stepsByOrder = getStepsByOrder(stepsLibrary);
@@ -18,19 +18,19 @@ export function useMultiStepNavigation(initialStep = Steps.PetName) {
   const isFirstStep = currentStepNumber === 0;
 
   const isLastStep = () => {
-    const totalSteps = Object.keys(Steps).length - 1;
+    const totalSteps = Object.keys(AdoptionSteps).length - 1;
 
     return currentStepNumber >= totalSteps;
   };
 
   const handleBack = () => {
-    const step = stepsByOrder[currentStepNumber - 1] as Steps;
+    const step = stepsByOrder[currentStepNumber - 1] as AdoptionSteps;
 
     setActiveStep(step);
   };
 
   const handleNext = () => {
-    const step = stepsByOrder[currentStepNumber + 1] as Steps;
+    const step = stepsByOrder[currentStepNumber + 1] as AdoptionSteps;
 
     setActiveStep(step);
   };
