+++ apps/mobile/src/components/register-adoption-form/compose/form-steps/index.ts
@@ -1,0 +1,1 @@
+export { default } from './form-steps'
