+++ apps/mobile/src/components/register-adoption-form/compose/stepper-controller/index.ts
@@ -1,0 +1,1 @@
+export { default } from './stepper-controller'
