+++ apps/mobile/src/components/register-adoption-form/compose/stepper-indicator/index.ts
@@ -1,0 +1,1 @@
+export { default } from './stepper-indicator'
