+++ apps/mobile/src/components/register-adoption-form/compose/stepper-indicator/stepper-indicator.tsx
@@ -1,0 +1,67 @@
+import React from "react";
+import { Box, Heading, Progress, Text, Image } from "native-base";
+import StepIcon from '../../../../../assets/step-background.png';
+import useLocale from "../../../../shared/hooks/use-locale";
+import { StepperIndicatorProps, AdoptionSteps } from "../../adoption-form.types";
+import theme from "../../../../theme";
+import { stepsLibrary } from "../../adoption-form.constants";
+
+function StepperIndicator({ activeStep }: StepperIndicatorProps) {
+  const { t } = useLocale();
+  
+  const totalSteps = Object.keys(stepsLibrary).length;
+  const stepNumber = stepsLibrary[activeStep].order + 1;
+  const processValue = (stepNumber * 100) / totalSteps;
+  const label = stepsLibrary[activeStep].label;
+
+  return (
+    <Box>
+      <Progress
+        position='absolute'
+        w='full'
+        value={processValue}
+        _filledTrack={{ rounded: 'none', borderBottomRightRadius: 'md' }}
+        rounded='none'
+        _ios={{
+          display: 'none'
+        }}
+      />
+      <Box
+        position='relative'
+        margin={8}
+        display='flex'
+        flexDirection='row'
+        justifyContent="space-between"
+        alignContent='center'
+        _ios={{
+          marginTop: "16"
+        }}
+      >
+        <Box>
+          <Text color={theme.colors.gray[600]}>{t("REGISTER_ADOPTION.TITLE")}</Text>
+          <Heading fontWeight='medium' color={theme.colors.primary[600]}>
+            {t(label)}
+          </Heading>
+        </Box>
+        <Image
+          position='absolute'
+          right={-18}
+          bottom={-14}
+          source={StepIcon}
+          alt="stepper icon"
+          size="md"
+        />
+        <Box alignItems="center" justifyContent="center">
+          <Text fontWeight="medium" color="white" fontSize='sm'>
+            {t('REGISTER_ADOPTION.STEP')}
+          </Text>
+          <Text lineHeight='sm' fontWeight="medium" color="white" fontSize='2xl'>
+            {stepNumber}/{totalSteps}
+          </Text>
+        </Box>
+      </Box>
+    </Box>
+  );
+};
+
+export default StepperIndicator;
